Created Tank Turret Rotation and TowerPawn Class
Turret rotation works by moving mouse left or right or using left and right arrow keys.

In the course they do:
Every frame they are tracking mouse location with a hit function in the world and then rotating the turret to face the mouse.

This means that it can spin really fast if moving the mouse over the tank and it's inconsistent to control because rotation is based relative cursor location tank in the middle. It also means that they had to put invisible walls around the map so the cursor always got a hit otherwise turret would not rotate.

I kept this method in the BasePawn so that the towers can use it to track the player, and added a parameter for interp speed. I also added a Doxygen comment so that when calling the method it says what it does.

For turret rotation I instead just check the X value of the mouse movement and add or subtract from the rotation value based on how fast the mouse is moving. Because of this, I can also then also keybind the rotation to the arrow keys. Arrow keys have a multiplication of 3 in the x axis so that it moves at a decent speed.

Camera is now in top down view, lighting adjusted to be more directly above